cmd/kube-proxy: refactor server.go
Split server.go into server.go and options.go and same for
server_test.go
